test(calendar): consolidate split recurrence-text test files
The two test files for src/common/utils/recurrence-text.ts lived in
different locations: the summary-helper tests in src/common/test/ and
the text-helper tests in src/common/test/utils/. Peer utility tests
(datePresets, ics-generator, instance-slug, render-markdown) all live
under src/common/test/utils/, so consolidate both suites into that
single conventional home. All 49 test cases are preserved verbatim;
no production code changes.
